requirements.txt
Django==2.1
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2019.3
